# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Question API swagger documentation",
        "contact": {
            "email": "echo.zahar@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.2"
    },
    "servers": [
        {
            "url": "/api/",
            "description": "Demo tech task API server"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "summary": "Регистрация нового пользователя",
                "description": "",
                "operationId": "e66f8a7452e13b03ae4e77dbb42e435c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRegister"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пользователь успешно зарегестрирован"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "summary": "Авторизация пользователя",
                "operationId": "d618a255475240ba3d40cc39692cd924",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful registration"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "summary": "logout user",
                "description": "Returns question data",
                "operationId": "Logged",
                "responses": {
                    "201": {
                        "description": "Successful registration"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthLoginRequest": {
                "title": "Авторизация.",
                "description": "Авторизация пользователей.",
                "properties": {
                    "email": {
                        "title": "email пользователя.",
                        "description": "email пользователя. Обязателеное поле, минимальная длина 5, максимальная 100. ",
                        "type": "string",
                        "format": "string",
                        "example": "customer@example.com"
                    },
                    "password": {
                        "title": "Пароль пользователя.",
                        "description": "введите пароль.",
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "AuthRegister": {
                "title": "регистрация нового пользователя.",
                "description": "Регистрация пользователей.",
                "properties": {
                    "name": {
                        "title": "имя пользователя",
                        "description": "Нужно ввести имя пользователя. Обязательное поле, минимальная длина 2, максимальное 50.",
                        "type": "string",
                        "format": "string",
                        "example": "Customer"
                    },
                    "email": {
                        "title": "email пользователя.",
                        "description": "email пользователя. Обязателеное поле, минимальная длина 5, максимальная 100. ",
                        "type": "string",
                        "format": "string",
                        "example": "customer@example.com"
                    },
                    "password": {
                        "title": "пароль пользователя.",
                        "description": "введите пароль",
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    },
                    "password_confirmation": {
                        "title": "подтверждение пароля",
                        "description": "введите пароль еще раз",
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "TOKEN": {
                "type": "apiKey",
                "description": "Этот контроллер создан исключительно для документаций и ни где не будет использоваться в проекте.",
                "name": "TOKEN",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}